refactor(playground): decompose PlaygroundApp - extract Tutorial/History/Toolbar controllers
